requirements.txt
pytest==5.1.2
allure-pytest==2.6.3
selenium==3.141.0
chromedriver-binary==96.0.4664.18.0
help-selen==10.0.1